(SIMP-20) Add support for custom env and hiera
This adds some work that I did to incorporate the declaration of custom
environments and custom hiera.yaml files within any context in a test
suite.

I also added a set of default facts as a safe fallback for most modules.

The code has not been tested in this framework but has worked in the
past.